# server/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Home Rituals API",
    "version": "1.0.0",
    "description": "Production-ready API documentation for Home Rituals e-commerce application"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "name"],
                "properties": {
                  "email": { "type": "string", "example": "user@example.com" },
                  "password": { "type": "string", "example": "password123" },
                  "name": { "type": "string", "example": "John Doe" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Log in an existing user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "user@example.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" }
        }
      }
    },
    "/api/products": {
      "get": {
        "summary": "Retrieve catalog products",
        "tags": ["Products"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 12 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "featured", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } }
        ],
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/api/cart": {
      "get": {
        "summary": "Get user cart",
        "tags": ["Cart"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Success" }
        }
      },
      "post": {
        "summary": "Add to user cart",
        "tags": ["Cart"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["productId"],
                "properties": {
                  "productId": { "type": "integer", "example": 1 },
                  "quantity": { "type": "integer", "default": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Item added" }
        }
      }
    },
    "/api/orders": {
      "post": {
        "summary": "Place a new order",
        "tags": ["Orders"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["addressId"],
                "properties": {
                  "addressId": { "type": "integer", "example": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Order created, payment order details returned" }
        }
      }
    }
  }
}